libm: Add optimized log2f
This implementation is based on code from ARM optimized routines [1].
The error for expf, for nearest rouding mode, is estimated on 0.82 ulp
with worst case.

It uses internal double precision arithmetics and a lookup table to
allow smaller polynomial and avoid the use of division.   The const
data is kept in a separate translation unit that allows fast code
for literal loads on most targets.

The performance shows about 2x better throughput and 1.5x better
latency for aarch64 (A72) using log2f benchmark.

[1] https://github.com/ARM-software/optimized-routines

Test: ran 32-bit and 64-bit x86 tests on host
Change-Id: I10aaefb88a0984bfe50c83176809748562a12f52
Signed-off-by: mydongistiny <jaysonedson@gmail.com>